dom/html/test/test_bug394700.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=394700
-->
<head>
  <title>Test for Bug 394700</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=394700">Mozilla Bug 394700</a>
<p id="display"></p>
<div id="content">
  <select><option id="A">A</option><option id="B">B</option></select>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 394700 **/

function remove(q1) { q1.remove(); }

function testSelectedIndex()
{
  document.addEventListener("DOMNodeRemoved", foo);
  remove(document.getElementById("B"));
  document.removeEventListener("DOMNodeRemoved", foo);

  function foo()
  {
    document.removeEventListener("DOMNodeRemoved", foo);
    remove(document.getElementById("A"));
  }
  var selectElement = document.getElementsByTagName("select")[0];
  is(selectElement.selectedIndex, -1, "Wrong selected index!");
  is(selectElement.length, 0, "Select shouldn't have any options!");
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(testSelectedIndex);
addLoadEvent(SimpleTest.finish);


</script>
</pre>
</body>
</html>
